backend/requirements.txt
flask
flask-cors
redis
Pillow == 9.3.0
numpy
pandas
ISR == 1.9.3
protobuf == 3.20.0
h5py==2.10.0